requirements.txt
annotated-types==0.7.0
async-timeout==5.0.1
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
greenlet==3.3.2
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-jose==3.5.0
redis==7.4.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
sqlmodel==0.0.37
typing-inspection==0.4.2
typing_extensions==4.15.0
